DSND-2874: Move consumer test package
