+++ src/test/java/uk/gov/companieshouse/exemptions/delta/kafka/consumer/ConsumerAspect.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.exemptions.delta.consumer;
+package uk.gov.companieshouse.exemptions.delta.kafka.consumer;
 
 import java.util.concurrent.CountDownLatch;
 import org.aspectj.lang.JoinPoint;
