+++ src/test/java/uk/gov/companieshouse/exemptions/delta/kafka/consumer/ConsumerNonRetryableExceptionTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.exemptions.delta.consumer;
+package uk.gov.companieshouse.exemptions.delta.kafka.consumer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -32,14 +32,14 @@ import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import uk.gov.companieshouse.delta.ChsDelta;
-import uk.gov.companieshouse.exemptions.delta.exception.RetryableException;
+import uk.gov.companieshouse.exemptions.delta.exception.NonRetryableException;
 import uk.gov.companieshouse.exemptions.delta.kafka.AbstractKafkaTest;
 import uk.gov.companieshouse.exemptions.delta.service.ServiceRouter;
 import uk.gov.companieshouse.exemptions.delta.kafka.KafkaUtils;
 
 @SpringBootTest
 @WireMockTest(httpPort = 8888)
-class ConsumerRetryableExceptionTest extends AbstractKafkaTest {
+class ConsumerNonRetryableExceptionTest extends AbstractKafkaTest {
 
     @Autowired
     private KafkaConsumer<String, byte[]> testConsumer;
@@ -54,7 +54,7 @@ class ConsumerRetryableExceptionTest extends AbstractKafkaTest {
     @DynamicPropertySource
     static void props(DynamicPropertyRegistry registry) {
         registry.add("spring.kafka.bootstrap-servers", confluentKafkaContainer::getBootstrapServers);
-        registry.add("steps", () -> 5);
+        registry.add("steps", () -> 1);
     }
 
     @BeforeEach
@@ -64,13 +64,13 @@ class ConsumerRetryableExceptionTest extends AbstractKafkaTest {
     }
 
     @Test
-    void testRepublishToErrorTopicThroughRetryTopics() throws InterruptedException, IOException {
+    void testRepublishToInvalidMessageTopicIfNonRetryableExceptionThrown() throws InterruptedException, IOException {
         //given
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         Encoder encoder = EncoderFactory.get().directBinaryEncoder(outputStream, null);
         DatumWriter<ChsDelta> writer = new ReflectDatumWriter<>(ChsDelta.class);
         writer.write(new ChsDelta("{}", 0, "context_id", false), encoder);
-        doThrow(RetryableException.class).when(router).route(any());
+        doThrow(NonRetryableException.class).when(router).route(any());
 
         //when
         testProducer.send(new ProducerRecord<>(COMPANY_EXEMPTIONS_DELTA_TOPIC, 0, System.currentTimeMillis(), "key", outputStream.toByteArray()));
@@ -79,10 +79,10 @@ class ConsumerRetryableExceptionTest extends AbstractKafkaTest {
         }
 
         //then
-        ConsumerRecords<?, ?> consumerRecords = KafkaTestUtils.getRecords(testConsumer, Duration.ofSeconds(10L), 6);
+        ConsumerRecords<?, ?> consumerRecords = KafkaTestUtils.getRecords(testConsumer, Duration.ofSeconds(10L), 2);
         Assertions.assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_TOPIC)).isOne();
-        assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_INVALID_TOPIC)).isZero();
-        assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_RETRY_TOPIC)).isEqualTo(4);
-        assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_ERROR_TOPIC)).isOne();
+        assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_INVALID_TOPIC)).isOne();
+        assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_RETRY_TOPIC)).isZero();
+        assertThat(KafkaUtils.noOfRecordsForTopic(consumerRecords, COMPANY_EXEMPTIONS_DELTA_ERROR_TOPIC)).isZero();
     }
 }
